+++ nodes/Autotask/helpers/http/index.ts
@@ -4,6 +4,13 @@ export * from './headers';
 // Import and export rate tracker initialization functions
 export * from './initRateTracker';
 
+// Import and export thread tracker initialization functions
+export * from './threadLimit';
+
 // Perform early initialization of the rate tracker
 import { initializeRateTrackerEarly } from './initRateTracker';
 initializeRateTrackerEarly();
+
+// Perform early initialization of the thread tracker
+import { initializeThreadTrackerEarly } from './threadLimit';
+initializeThreadTrackerEarly();